main/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
Flask
python-dotenv
flask-cors

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

azure-storage-blob
azure-identity